// src/shared/api/models/postApiStockBodyType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type PostApiStockBodyType = (typeof PostApiStockBodyType)[keyof typeof PostApiStockBodyType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostApiStockBodyType = {
    DAMAGED: 'DAMAGED',
    LOST: 'LOST',
    'RE-COUNTED': 'RE-COUNTED',
    RECEIVED: 'RECEIVED',
    REFUNDED: 'REFUNDED',
    RESTOCKED: 'RESTOCKED',
    THEFT: 'THEFT',
    EXPIRED: 'EXPIRED',
    TRANSFERRED: 'TRANSFERRED',
    RETURNED: 'RETURNED',
    SOLD: 'SOLD',
    'RETURNED-TO-SUPPLIER': 'RETURNED-TO-SUPPLIER',
} as const;